+++ RTSProject/RTSProject/Effect.cpp
@@ -184,9 +184,9 @@ void EffectFireBall::Render(std::shared_ptr<Camera> camera)
 							mTransMat.mScale * 0.2f };
 
 	TransformMat tranMat;
-	for (int t = 0; t < 5; t++)
+	for (size_t t = 0; t < 5; t++)
 	{
-		for (int i = 0; i < 4; ++i)
+		for (size_t i = 0; i < 4; ++i)
 		{
 			tranMat.mTrans = positions[t];
 			tranMat.mRot = mTransMat.mRot + rotations[i];
