+++ RTSProject/RTSProject/UnitStateGotoObj.cpp
@@ -44,7 +44,7 @@ bool UnitStateGotoObj::FindEnemyInRange(Unit * ownUnit, std::shared_ptr<Terrain>
 	float minRange = 100.0f;
 	glm::vec3 pos2 = ownUnit->GetPosition();
 
-	for (int i = 0; i < mEnemy.size(); ++i)
+	for (size_t i = 0; i < mEnemy.size(); ++i)
 	{
 		shared_ptr<RTSObject> enemyUnit = mEnemy[i].lock();
 		if (enemyUnit)
